backend/requirements.txt
pip>=20.1.1

Django>=3.0.7
psycopg2-binary>=2.8.5
gunicorn>=20.0.4

django-filter>=2.2.0
django-cors-headers>=2.4.0
djangorestframework>=3.11.0
djangorestframework_simplejwt>=4.3.0

elasticsearch-dsl==7.1.0
django-elasticsearch-dsl==7.1.1.
django-elasticsearch-dsl-drf>=0.20.5
drf-yasg


# Developer Tools
ipdb>=0.11
ipython>=6.4.0
flake8>=3.7.7

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose>=1.4.5
nose-progressive>=1.5.1
coverage>=4.5.1

# reading environment variables
python-dotenv

# logging
json-logging>=0.1.1
coloredlogs>=10.0

requests

